Creates a common/ folder and splits utils.py into different sub-scripts: - utils.py general functions - scrape_utils.py functions to support scraping the BOE site for minutes - parse_utils.py functions to support parsing downloaded pdfs
