+++ common/parse_utils.py
@@ -1,47 +1,38 @@
-from PyPDF2 import PdfFileReader
+import PyPDF2
 import datetime as dt
+import pandas as pd
+import re
+from datetime import datetime
+
+from common.utils import replace_chars
 
 REPLACEMENTS = [
-    ('Œ', '-'),
-    ('ﬁ', '"'),
-    ('ﬂ', '"'),
-    ('™', "'"),
-    ('Ł', '•'),
-    # (',', "'"),
-    ('Š', '-'),
-    ('€', ' '),
-    ('¬', '-'),
-    ('–', '…'),
-    ('‚', "'"),
-    ('Ž', '™'),
-    ('˚', 'fl'),
-    ('˜', 'fi'),
-    ('˛', 'ff'),
-    ('˝', 'ffi'),
-    ('š', '—'),
-    ('ü', 'ti'),
-    ('î', 'í'),
-    ('è', 'c'),
-    ('ë', 'e'),
-    ('Ð', '–'),
-    ('Ò', '"'),
-    ('Ó', '"'),
-    ('Õ', "'"),
+    ("Œ", "-"),
+    ("ﬁ", '"'),
+    ("ﬂ", '"'),
+    ("™", "'"),
+    ("Ł", "•"),
+    ("Š", "-"),
+    ("€", " "),
+    ("¬", "-"),
+    ("–", "…"),
+    ("‚", "'"),
+    ("Ž", "™"),
+    ("˚", "fl"),
+    ("˜", "fi"),
+    ("˛", "ff"),
+    ("˝", "ffi"),
+    ("š", "—"),
+    ("ü", "ti"),
+    ("î", "í"),
+    ("è", "c"),
+    ("ë", "e"),
+    ("Ð", "–"),
+    ("Ò", '"'),
+    ("Ó", '"'),
+    ("Õ", "'"),
 ]
 
-def replace_chars(text, replacement_list):
-    """Replaces a set of characters specified by a list of replacement keys
-
-    Args:
-        text (str): The raw text whose characters will be replaced
-        replacement_list (list): List of tuples in which the first item
-        is the current character and the second item is its replacement
-    Returns:
-        text (str): The text with the characters replaced
-    """
-    for current, new in replacement_list:
-        text = text.replace(current, new)
-    return text
 
 def parse_pdf(pdf_path):
     """Parses the pdf of the minutes from a BOE meeting and cleans the text
@@ -59,6 +50,46 @@ def parse_pdf(pdf_path):
         print(f"The following error occurred parsing file '{pdf_path}': {e}")
         raise e
     return minutes
+
+
+def store_pdf_text_to_df(path):
+    """Finds .pdf files stored at the given url and stores them within the
+    repository for later analysis.
+
+    Args:
+        base_url (str): The main url for the Comptroller of Baltimore's webiste
+        minutes_url (str): The url where the function can find links to pages of
+        pdf files organized by year
+    Returns:
+        None: This is a void function.
+    """
+    pdf_paths = list(path.rglob("*.pdf"))
+    text_df = pd.DataFrame(columns=["date", "page_number", "minutes"])
+    for pdf_path in pdf_paths:
+        # print(f"Parsing file: {pdf_path.name}")
+        minutes = ""
+        pdfFileObj = open(pdf_path, "rb")
+        try:
+            pdfReader = PyPDF2.PdfFileReader(pdfFileObj, strict=False)
+        except ValueError:
+            print(f"An error occurred reading file {pdf_path}")
+        for page in pdfReader.pages:
+            minutes += page.extractText().strip()
+
+        date_string = pdf_path.stem
+        date = datetime.strptime(date_string, "%Y_%m_%d").date()
+        page_number = re.findall(r"(^[0-9]+)", minutes)
+        if page_number:
+            page_number = page_number[0]
+        else:
+            page_number = ""
+        try:
+            row = {"date": date, "page_number": page_number, "minutes": minutes.strip()}
+            text_df = text_df.append(row, ignore_index=True)
+        except ValueError:
+            print(f"No date found for file {pdf_path}")
+    print(f"Wrote {len(text_df)} rows to the table of minutes.")
+    return text_df
 
 
 class Minutes:
@@ -87,9 +118,8 @@ class Minutes:
             that will be stored in self.reader
         """
         file = open(pdf_path, "rb")
-        reader = PdfFileReader(file, strict=False)
+        reader = PyPDF2.PdfFileReader(file, strict=False)
         return reader
-
 
     def parse_date(self, pdf_path):
         """Parses a datetime object from the path to the pdf file for use
@@ -106,7 +136,6 @@ class Minutes:
         date_str = pdf_path.stem
         date = dt.datetime.strptime(date_str, "%Y_%m_%d")
         return date
-
 
     def parse_pages(self):
         """Extracts text from pdf pages and stores it in self.raw_text
@@ -133,7 +162,7 @@ class Minutes:
             clean_text (str): Returns the text parsed from the pdf pages as a
             string and also stores that text in self.clean_text
         """
-        clean_text = " ".join(self.raw_text.split()) # remove double spaces
+        clean_text = " ".join(self.raw_text.split())  # remove double spaces
         clean_text = replace_chars(clean_text, REPLACEMENTS)
         self.clean_text = clean_text
         return clean_text
